\name{mygbif-package}
\alias{mygbif-package}
\alias{mygbif}
\docType{package}
\title{
Package has single functionality of getting count of records for a taxon on GBIF
~~ mygbif ~~
}
\description{
Package has single functionality of getting count of records for a taxon on GBIF
}
\details{
\tabular{ll}{
Package: \tab mygbif\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-04-11\cr
License: \tab What license is it under?\cr
}
Use to get count of records
}
\author{
Vijay Barve
Maintainer: Vijay Barve <vijay.barve@gmail.com>
}
